# backend/api-docs.yaml
openapi: 3.0.0
servers: []
info:
  version: "1.0.0"
  title: KFC SALIDA DE EQUIPOS
  description: API PARA SALIDA DE EQUIPOS
paths:
...
  /salida_equipos/v1/adjuntos:
    post:
      tags:
        - Adjuntos
      summary: Crea Adjunto para un Detalle de Salida
      security:
        - bearerAuth: []
      description: Crea Adjunto
      operationId: crearAdjunto
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                salidaDetalleId:
                  type: integer
                  format: int32
                  example: 1
                archivos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Adjunto creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: Adjunto creado
                  datos:
                    $ref: '#/components/schemas/AdjuntosModel'
                  codigo:
                    type: integer
                    format: int32
                    example: 201
                  exito:
                    type: boolean
                    example: true
        '400':
          description: Error al crear adjunto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        codigo:
          type: integer
          format: int32
          example: 400
        mensaje:
          type: string
          example: Error
        datos:
          type: object
          example: null
        exito:
          type: boolean
          example: false
...
    AdjuntosModel:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1        
        salidaDetalleId:
          type: integer
          format: int32
          example: 1
        data:
          type: string
          example: 'ENCRYPTEDDATAFORIMAGE'
        tipo:
          type: string
          example: IMAGEN
        extension:
          type: string
          example: image/png        
        createdAt:
          type: string
          format: date-time
          example: 2025-01-02T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-01-02T00:00:00Z
...